Ticket #103: Add Win64 Support to the FDO API and SDF Provider
git-svn-id: http://svn.osgeo.org/fdo/trunk@3189 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Fdo/Unmanaged/Src/Common/Xml/ReaderXrcs.cpp
@@ -399,7 +399,7 @@ unsigned int FdoXmlReaderXrcs::InputStream::readBytes(
     const unsigned int  maxToRead 
 )
 {
-    return( mStream->Read( toFill, maxToRead ) );
+    return (unsigned int)(( mStream->Read( toFill, maxToRead ) ));
 }
